Moved to using inheritance for voxel meshes
